feat: move headers into own module
